requirements.txt
streamlit
requests
python-dotenv
python-docx
beautifulsoup4
